requirements.txt
fastapi
uvicorn
fastai
pillow
scikit-image
openpyxl
pandas
